tracking details clean up and refactor of logic for seperation
