+++ package.json
@@ -11,6 +11,7 @@
         "@types/react": "^18.0.0",
         "@types/react-dom": "^18.0.0",
         "bootstrap": "^5.3.0",
+        "geolib": "^3.3.4",
         "mapbox-gl": "^2.15.0",
         "react": "^18.2.0",
         "react-bootstrap": "^2.7.4",
@@ -48,6 +49,7 @@
         ]
     },
     "devDependencies": {
+        "@types/formatcoords": "^1.1.0",
         "@types/mapbox-gl": "^2.7.11"
     }
 }
