+++ src/Apis/DecimalCoordsToDeg.tsx
@@ -1,0 +1,8 @@
+import { decimalToSexagesimal } from "geolib";
+
+export const DecimalCoordsToDeg = (lat: number, lng: number) => {
+    const coors = [];
+    coors.push(decimalToSexagesimal(lat));
+    coors.push(decimalToSexagesimal(lng));
+    return coors;
+};
