+++ src/Components/SideBar/TrackingDetails/ThreatLevelDropdown/ThreatLevelDropdownS.scss
@@ -1,0 +1,14 @@
+.ThreatLevel.dropdown {
+    display: flex;
+    width: 40rem;
+
+    .btn {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        width: 30rem;
+        height: 3rem;
+        padding-bottom: 0;
+        border-radius: 0;
+    }
+}
