+++ src/Components/SideBar/TrackingDetails/TrackingDetailsHeader/TrackingDetailsHeaderS.scss
@@ -1,0 +1,22 @@
+.DetailsHeader {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+
+    .name {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+
+        width: 100%;
+    }
+
+    .coors {
+        display: flex;
+        flex-direction: row;
+    }
+
+    .lat {
+        padding-right: 0.5rem;
+    }
+}
